luluisher/requirements.txt
sDjango==4.0.4
django-cors-headers==3.12.0
PyMySQL==1.0.2
